fix: convert new modules to new format
