+++ tsconfig.browser.json
@@ -5,12 +5,12 @@
         "noEmit": true,
         "types": [
             "vite/client"
-        ],
+        ]
     },
     "include": [
         "src/browser/**/*.tsx",
         "src/browser/**/*.ts",
         "src/types/**/*.ts",
         "src/types/augment-window.d.ts"
-    ],
+    ]
 }
